sanitize-data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.7.0